+++ src/components/Inputs/utils.ts
@@ -1,7 +1,7 @@
 import type { Control as OriginalControl, FieldPath, FieldValues } from 'react-hook-form';
 
 
-/** react-hook-form's Control type would make some components throw errors. Simplified type. */
+/** Simplified type of react-hook-form's Control. It would make some components throw errors, like 'type excessively deep'. */
 export type Control<F extends FieldValues = FieldValues> = Pick<OriginalControl<F>, '_formValues'>;
 export type ControlIds<C extends Control> = FieldPath<C['_formValues']>;